Hold to dismiss, for #20
Instead of dismissing instantly on touch, instead the touch must be held
down for 2 seconds. During this time, the background will fade to be the
same color as the alarm icon, giving a progress indicator.

While holding, the icon will also change to a finger (touch_app), to
provide a hint about what you need to do. This is an effective hint
because tapping makes it flicker, and if you hold for long enough to see
what the new icon is, you'll also hold for long enough that the
background will start fading, giving you the next hint.

Pressing the home or back button will still dismiss instantly, since
there's no way to track touches on those.